+++ db/bankDB.py
@@ -26,7 +26,7 @@ class Bank:
                     #"credit_score": 0
                 }
                 update_scheduler(lambda: cache_initiator.add_to_user_cache(user_id, bank_data=user))
-                request_threading(lambda: bank_collection.insert_one(user)).result()
+                request_threading(lambda: bank_collection.insert_one(user), user_id).result()
                 logger.info(f"User {user_id} has been created successfully.")
         except Exception as e:
             logger.error("Error encountered while creating the user.", e)
@@ -39,7 +39,7 @@ class Bank:
             user_data = request_threading(lambda: bank_collection.find_one({"user_id": user_id})).result()
             if user_data:
                 update_scheduler(lambda: cache_initiator.delete_user_cache(user_id))
-                request_threading(lambda: bank_collection.delete_one({"user_id": user_id})).result()
+                request_threading(lambda: bank_collection.delete_one({"user_id": user_id}), user_id).result()
                 logger.info("User has been deleted successfully.")
             else:
                 logger.warning("User not found.")
@@ -55,7 +55,7 @@ class Bank:
             if user_data:
                 user_data['bank'] = points
                 update_scheduler(lambda: cache_initiator.update_user_cache(user_id, bank_data=user_data))
-                request_threading(lambda: bank_collection.update_one({"user_id": user_id}, {"$set": {"bank": points}})).result()
+                request_threading(lambda: bank_collection.update_one({"user_id": user_id}, {"$set": {"bank": points}}), user_id).result()
                 logger.info("User has been updated successfully.")
             else:
                 logger.warning("User not found.")
@@ -71,7 +71,7 @@ class Bank:
             if user_data:
                 user_data['upgrades'] = upgrades
                 update_scheduler(lambda: cache_initiator.update_user_cache(user_id, bank_data=user_data))
-                request_threading(lambda: bank_collection.update_one({"user_id": user_id}, {"$set": {"upgrades": upgrades}})).result()
+                request_threading(lambda: bank_collection.update_one({"user_id": user_id}, {"$set": {"upgrades": upgrades}}), user_id).result()
                 logger.info("User has been updated successfully.")
             else:
                 logger.warning("User not found.")
